Improve loading of internal state
The previous strategy was prone to error, as the state could exist while
being unloaded. It's now in an Option to leverage the type system to
avoid errors.

This also enables commands that do not need the internal state to work
even though the internal state is broken (necessary for factory reset
notably)