Fix docker image
This uses an arch linux docker image because:

Debian 12 does not have gnupg 2.4
Ubuntu 24.04 has it and vsmartcard but not cargo-tarpaulin. And building cargo-tarpaulin seems to always lead to a dynamic dependency on `libssl1.1.so` which is not available in the repos.

This is still not perfect. Building vsmartcard fails, so I'm just bundling a binary package...